Renderer is leftover when we are doing a flip from display:contents to display:none and then calling focus() https://bugs.webkit.org/show_bug.cgi?id=250354 rdar://102808896
Reviewed by NOBODY (OOPS!).

The basic bug here is that we are doing a flip from display:contents to display:none and then calling focus(). That computes focusability using Element::resolveComputedStyle and that causes us to forget we used to have display:contents . Then we force render tree update which fails to remove the text renderer in the formerly display:contents tree because it thinks it is a display:none tree without renderers. The fix updates a flag in the parentNode of the left over renderer to mark that a renderer is created for a child by its anscestor.

* Source/WebCore/dom/ContainerNode.cpp:
(WebCore::childNodesMayHaveRenderers):
(WebCore::destroyRenderTreeIfNeeded):
* Source/WebCore/dom/Node.cpp:
(WebCore::Node::setAncestorCreatedRendererForChildNodes const):
(WebCore::Node::ancestorCreatedRendererForChildNodes const):
* Source/WebCore/dom/Node.h:
* Source/WebCore/rendering/updating/RenderTreeUpdater.cpp:
(WebCore::RenderTreeUpdater::updateTextRenderer):