[GLib] Remove download-started vfunc from WebKitWebContextClass https://bugs.webkit.org/show_bug.cgi?id=251521
Reviewed by Adrian Perez de Castro.

Now in theory, this class struct should become private and move to the
source file, because WebKitWebContext is no longer derivable and there
is no point to having a public vtable anymore because nothing can use
it.

But in the meantime, this signal is gone and shouldn't be here.

* Source/WebKit/UIProcess/API/glib/WebKitWebContext.h.in:

Canonical link: https://commits.webkit.org/259727@main