[css-animations] composite operation of implicit keyframes for CSS Animations should be "replace" https://bugs.webkit.org/show_bug.cgi?id=251499
Reviewed by Antti Koivisto.

The CSS Animations spec indicates in its "Keyframes" section [0] that the default composite is
"replace". As such, when generating an implicit keyframe for an effect tied to a CSS Animation,
we should set the composite value to "replace".

This issue was caught by a new version of css/css-animations/KeyframeEffect-getKeyframes.tentative.html
which was not yet in our repository, so we update it with the most recent changes.

We make an additional change in the final subtest of this WPT to correctly have "replace" as
the composite value for the implicit keyframes. I suspect this change was not made when Google
last changed that test [1] because Chrome fails to generate the right implicit keyframes in this case.

[0] https://drafts.csswg.org/css-animations-2/#keyframes
[1] https://github.com/web-platform-tests/wpt/commit/1a5c61d2ca9d0f2c2801adf2b433c7931512152f

* LayoutTests/imported/w3c/web-platform-tests/css/css-animations/KeyframeEffect-getKeyframes.tentative.html:
* Source/WebCore/rendering/style/KeyframeList.cpp:
(WebCore::KeyframeList::fillImplicitKeyframes):

Canonical link: https://commits.webkit.org/259739@main