[WGSL] Allow early compilation without hints https://bugs.webkit.org/show_bug.cgi?id=251494 rdar://104908379
Reviewed by Myles C. Maxfield.

This is just a temporary workaround so we can start passing some CTS tests.
For now, since we are not yet using the layout hints, we can just try to
compile upfront even without them. Later we'll wire this up correctly so that
if no hints are provide the compilation is deferred.

* Source/WebGPU/WebGPU/ShaderModule.mm:
(WebGPU::Device::createShaderModule):

Canonical link: https://commits.webkit.org/259747@main