[web-animations] animation of color list custom properties with iterationComposite is incorrect https://bugs.webkit.org/show_bug.cgi?id=251574
Reviewed by Antti Koivisto.

Certain types require interpolation for each iteration when `iterationComposite` is used. While we
had logic in place for this in CSSPropertyAnimation::propertyRequiresBlendingForAccumulativeIteration()
for `SyntaxValue` custom properties, we did not for `SyntaxValueList`. We factor the code that determined
whether interpolation was required for the `SyntaxValue` case in a new static method which we now also
call for each value in a `SyntaxValueList`.

* LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/animation/custom-property-animation-color-comma-list-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/animation/custom-property-animation-color-space-list-expected.txt:
* Source/WebCore/animation/CSSPropertyAnimation.cpp:
(WebCore::syntaxValuesRequireBlendingForAccumulativeIteration):
(WebCore::CSSPropertyAnimation::propertyRequiresBlendingForAccumulativeIteration):

Canonical link: https://commits.webkit.org/259761@main