Regression(LazyIframeLoading): "MySize ASSIST" view is empty in UNIQLO app https://bugs.webkit.org/show_bug.cgi?id=251639 rdar://104519846
Reviewed by Ryosuke Niwa.

The "MySize ASSIST" view contains an iframe that is initially outside the
viewport and marked as `loading="lazy"`. There is JS logic on the page to wait
for the iframe's load event before moving the iframe to the viewport.
However, since the iframe is outside the viewport and loads lazily, its load
event never fires and the view remains blank.

This looks like a content issue so I am adding a quirk for the UNIQLO app
which deactivates lazy iframe loading with a linked-on-after SDK check.

* Source/WTF/wtf/cocoa/RuntimeApplicationChecksCocoa.h:
* Source/WebCore/html/HTMLIFrameElement.cpp:
(WebCore::HTMLIFrameElement::shouldLoadFrameLazily):
* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::shouldDisableLazyIframeLoadingQuirk const):
* Source/WebCore/page/Quirks.h:
* Source/WebCore/platform/RuntimeApplicationChecks.h:
* Source/WebCore/platform/cocoa/RuntimeApplicationChecksCocoa.mm:
(WebCore::IOSApplication::isUNIQLOApp):

Canonical link: https://commits.webkit.org/259790@main