[JSC] Disable import-assertion since TC39 decides bringing it back to stage-2 https://bugs.webkit.org/show_bug.cgi?id=251600 rdar://104964116
Reviewed by Michael Saboff.

We found that the current syntax & semantics do not work with CSS / JSON module integration in HTML side,
and it requires drastic changes in import-assertion in both syntax and semantics levels, which means that
the current proposal becomes obsolete / not-shippable state. TC39 discussed about this and decided to
bring import-assertion down to stage-2 from stage-3 to announce that this is not shippable state right now.
This patch disables import-assertion by flipping a runtime flag.

* Source/JavaScriptCore/runtime/OptionsList.h:

Canonical link: https://commits.webkit.org/259795@main