Some data task signposts are not ended properly https://bugs.webkit.org/show_bug.cgi?id=251630 rdar://104976161
Reviewed by Alex Christensen.

Sometimes, when a data task is cancelled, the didCompleteWithError method is not called. Since that
is where we end data task signposts, this can lead to dropped signposts and incorrectly attributed
resource loads in perf traces.

To fix this, move the end signpost call to the destructor instead. This matches the begin signpost
call, which occurs in the constructor.

Note that this only affects behavior when a perf tracing tool is explicitly enabled; these signposts
are disabled otherwise.

* Source/WebKit/NetworkProcess/cocoa/NetworkDataTaskCocoa.mm:
(WebKit::NetworkDataTaskCocoa::NetworkDataTaskCocoa):
(WebKit::NetworkDataTaskCocoa::~NetworkDataTaskCocoa):
(WebKit::NetworkDataTaskCocoa::didCompleteWithError):

Canonical link: https://commits.webkit.org/259826@main