Remove unnecessary masking after wasm LLInt comparison instructions. https://bugs.webkit.org/show_bug.cgi?id=251769 rdar://105063662
Reviewed by Tadeu Zagallo.

Currently, we always follow a LLInt comparison instructions with an `and with 1` mask.
This is completely unnecessary because the LLInt comparison instructions already always
produce a 1 or a 0.  Hence, the `and with 1` mask is effectively a no-op.  This patch
removes this unnecessary masking.

* Source/JavaScriptCore/llint/WebAssembly.asm:
* Source/JavaScriptCore/llint/WebAssembly32_64.asm:
* Source/JavaScriptCore/llint/WebAssembly64.asm:

Canonical link: https://commits.webkit.org/259882@main