Pin wasmMemoryBasePointer on BBQ/ARM https://bugs.webkit.org/show_bug.cgi?id=251845
Reviewed by NOBODY (OOPS!).

We have enough registers to pin the wasm base memory value (but not the bounds
checking size) on BBQ for ARM. We don't have enough registers to do the same on
llint. This also means that we need to be careful about loading
wasmBaseMemoryPointer; t7 can be used as a temporary at virtually any point in
the llint code, so we call reloadMemoryRegistersFromInstance as the last thing
we do before switching to the BBQ code.

This results in a 0.7% (statistically significant) improvement on the wasm parts
of JetStream2 (i.e. wasm-cli.js) on ARM.

* Source/JavaScriptCore/jit/GPRInfo.h:
* Source/JavaScriptCore/llint/WebAssembly.asm:
* Source/JavaScriptCore/wasm/WasmAirIRGenerator32_64.cpp:
(JSC::Wasm::AirIRGenerator32::emitCheckAndPreparePointer):
* Source/JavaScriptCore/wasm/WasmAirIRGenerator64.cpp:
* Source/JavaScriptCore/wasm/WasmAirIRGeneratorBase.h:
(JSC::Wasm::ExpressionType>::AirIRGeneratorBase):
(JSC::Wasm::ExpressionType>::restoreWebAssemblyGlobalState):
(JSC::Wasm::ExpressionType>::emitIndirectCall):
* Source/JavaScriptCore/wasm/WasmBinding.cpp:
(JSC::Wasm::wasmToWasm):
* Source/JavaScriptCore/wasm/js/JSToWasm.cpp:
(JSC::Wasm::createJSToWasmWrapper):
* Source/JavaScriptCore/wasm/js/WebAssemblyFunction.cpp:
(JSC::WebAssemblyFunction::jsCallEntrypointSlow):