Fix statement depth for parsing static block https://bugs.webkit.org/show_bug.cgi?id=251760 rdar://105104494
Reviewed by Mark Lam.

Static block is treated as a function in spec, that is it should be
parsed as a function and executed as a function invocation. However,
the implementation of parsing class static block is based on
parseBlockStatement. To be consistent with the function environment
during parsing, we should initialize the statement depth to 0.

* JSTests/stress/function-lexical-declaration.js: Added.
(A.C):
(A):
* Source/JavaScriptCore/parser/Parser.h:
(JSC::Scope::declareFunction):

Canonical link: https://commits.webkit.org/259981@main