Replace deprecated CCCryptorGCM function where possible https://bugs.webkit.org/show_bug.cgi?id=251851 <rdar://problem/30660074>
Reviewed by NOBODY (OOPS!).

We can use CCCryptorGCMOneshotEncrypt and CCCryptorGCMOneshotDecrypt
instead.

*Source\WebCore\PAL\pal\spi\cocoa\CommonCryptoSPI.h:(CCCryptorGCM):
Deleted.
(CCCryptorGCMOneshotEncrypt): Added.

*Source\WebCore\crypto\mac\CryptoAlgorithmAES_GCMMac.cpp:(encryptAES_GCM): Replaced
CCCryptorGCM with CCCryptorGCMOneshotEncrypt.
(decyptAES_GCM): Replaced CCCryptorGCM with CCCryptorGCMOneshotDecrypt.

Note that CCCryptorGCMOneshotDecrypt does not require the use of
constantTimeMemcmp afterwards because that is part of the
CCCryptorGCMOneshotDecrypt function, which means if the comparison
fails, the function returns an OperationError exception.